Reworked inflow builder interface. Ensure that refresh().await() returns data from the cache, not from the loader.

+++ library/core/src/main/kotlin/inflow/Progress.kt
@@ -21,16 +21,16 @@ sealed class Progress {
 
     /**
      * Loading progress ([current] / [total]) as tracked by [ProgressTracker] instance passed to
-     * [InflowConfig.loader].
+     * the loader (see [InflowConfig.data]).
      *
      * Percentage can be calculated with [rate] function.
      */
-    data class State internal constructor(val current: Float, val total: Float) : Progress() {
+    data class State internal constructor(val current: Double, val total: Double) : Progress() {
         /**
          * Percentage calculated as `current / total` (but never greater than `1`).
          * If `total <= 0` or `current < 0` then `0` will be returned.
          */
-        fun rate(): Float = if (total > 0f && current >= 0f) min(current / total, 1f) else 0f
+        fun rate(): Double = if (total > 0.0 && current >= 0.0) min(current / total, 1.0) else 0.0
     }
 }
 
@@ -38,6 +38,5 @@ sealed class Progress {
  * Intermediate loading progress tracker.
  */
 interface ProgressTracker {
-    fun state(current: Float, total: Float)
+    fun state(current: Double, total: Double)
 }
-
